refactor(core): relocate endpoint cluster out of list framework
Move the live, now-ngrx-free endpoint cells + helper (endpoint-card,
table-cell-endpoint-address/name/status, endpoint-list.helpers incl. the
public-API EndpointListDetailsComponent base) from
shared/components/list/list-types/endpoint into a dedicated
shared/components/endpoint-list so workstream D can delete the list
framework. Repoint TableCellCustom/CardCell at signal-list/cell-base,
drop the vestigial listCards self-registration, and update public-api,
SharedModule, and the core/kube/git/metrics consumers. The dead
table-cell-endpoint-details + endpoint-data-source.helpers stay behind
for deletion in D.
