+++ src/frontend/packages/core/src/shared/components/endpoint-list/endpoint-card/endpoint-card.component.scss
