+++ src/frontend/packages/core/src/shared/components/endpoint-list/table-cell-endpoint-address/table-cell-endpoint-address.component.html
