+++ src/frontend/packages/core/src/shared/components/endpoint-list/table-cell-endpoint-status/table-cell-endpoint-status.component.html
