+++ src/frontend/packages/core/src/shared/components/endpoint-list/table-cell-endpoint-status/table-cell-endpoint-status.component.spec.ts
@@ -3,7 +3,7 @@ import { provideZonelessChangeDetection } from '@angular/core';
 import { describe, it, expect, beforeEach, afterEach, beforeAll, afterAll, vi } from 'vitest';
 import { EndpointModel } from '@stratosui/store';
 import { createBasicStoreModule, CoreTestingModule } from '@test-framework';
-import { CoreModule } from '../../../../../../core/core.module';
+import { CoreModule } from '../../../../core/core.module';
 import { TableCellEndpointStatusComponent } from './table-cell-endpoint-status.component';
 
 describe('TableCellEndpointStatusComponent', () => {
