+++ src/frontend/packages/core/src/shared/components/list/list-types/endpoint/table-cell-endpoint-details/table-cell-endpoint-details.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { describe, it, expect, beforeEach } from 'vitest';
 import { BaseTestModules } from '@test-framework/core-test.helper';
-import { EndpointListHelper } from '../endpoint-list.helpers';
+import { EndpointListHelper } from '../../../../endpoint-list/endpoint-list.helpers';
 import { TableCellEndpointDetailsComponent } from './table-cell-endpoint-details.component';
 
 describe('TableCellEndpointDetailsComponent', () => {
